Tabs: Ids, classes, and building blocks (#10)
* Tabs Enhancements

- allow the tabs class to be passed in (previously defaulted to `is-boxed`, so this is a change)
- wrap the contents in a div and allow a class to be passed in
- add content to the right of the tabs, such as a button or display

* Table: Model name from query instead of first record

Reading the first record caused an extra query. This also works whether or not there are rows.

* Add ids for containers in Tabs component

* Building blocks for Tabs

This breaks out two new components, one for a tab and one for its content. This allows tabs and content to be created outside the scope of the Tabs component.

This might be useful for adding a tab and content section via a Turbo response.

* Assign Id to immediate tabs container

This moves it from the div to the ul.

* Add namespace to SimulusDataAttributes