Eliminate load order dependency for phlex-rails
Checking for Phlex::Rails in the code introduced a load order dependency: if the phlex-rails gem was below this gem in the Gemfile, it would not be seen and the additional functionality would not be included.

This change checks for Rails at startup, then defers checks for Phlex::Rails (and Turbo) to an initializer in a Railtie.