Release 0.7.0: Navigation Bar Changes
This introduces the ability to specify additional classes for the `nav` element as well optionally wrap the content with a container.

This release changes the default behavior of the navigation bar. Previously, it added classes ("is-light block") to the `nav` element and wrapped everything underneath the `nav` element with a container (`<div class="container">`).

To match the prior output, update the call to the component as follows:

    Bulma::NavigationBar(classes: "is-light block", container: true)